site/newsletter/pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mlsysbook-news"
version = "0.1.0"
description = "CLI for the ML Systems newsletter. Drafts in git, send via Buttondown."
readme = "cli/README.md"
requires-python = ">=3.10"
authors = [
    { name = "Vijay Janapa Reddi", email = "vj@eecs.harvard.edu" },
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
]

# Runtime dependencies for the news CLI.
# Pinned to major-version floors that expose the APIs we rely on,
# with upper bounds loose enough to follow upstream minor releases.
dependencies = [
    "rich>=13.0.0",                 # colored terminal output, tables, panels
    "requests>=2.31.0",             # Buttondown REST client
    "python-frontmatter>=1.1.0",    # parse YAML frontmatter in markdown drafts
]

[project.scripts]
# `pip install -e .` from site/newsletter/ registers `news` on PATH.
news = "cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.6.0",
]

[project.urls]
Homepage = "https://mlsysbook.ai/newsletter/"
Repository = "https://github.com/harvard-edge/cs249r_book"
Issues = "https://github.com/harvard-edge/cs249r_book/issues"

[tool.setuptools]
packages = ["cli", "cli.core", "cli.commands"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
